requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
python-multipart
pydantic
cryptography  # Required for pymysql
